See description
PJ 2.6 Changes
- LifeTimerAnimation renamed to AnimatedLifeTimer
- FlashTimer now starts with the flash ON instead of flash OFF (makes more sense)
- RotateAnimation2D supports GameObjects without a Node2D component
- Formation allows you to specify “spawns locations” as children of the formation, which are used to populate the formation from an object list
- LinePathLayout: Layout child objects along a line
- PathLayouts now support orient to path
- Fixed some layout bugs
- SomeGoal, IntGoal model objects for quests
- Stage2D/StageEntity: used for slide in/slide out animations. A StageEntity is a child of a Stage2D, it will be given a default Enter/Exit path, but you can supply your own. This is useful for Visual Novels
- ScreenMeshSizer2D: size a mesh to the screen size
- ScreenTransformLimiter2D: limit the bounds of a child object to the screen
- WeaponEmitter2D now supports multiple spawned bullet objects with weights and a maxAlive counter to prevent more than N bullets from being alive in gamespace simultaneously

+++ PhaseJumpPro/Assets/phasejumppro/Core/ObjectListBuilder.cs
@@ -48,7 +48,7 @@ namespace PJ
 				switch (method)
                 {
 					case Method.Random:
-						chooseIndex = Mathf.RoundToInt(random.Value * (objectList.Count - 1));
+						chooseIndex = Mathf.RoundToInt(random.Value * (float)(objectList.Count - 1));
 						break;
                 }
